reorganise journey mock data
