+++ playwright/mockData/orgs/KPD/journeys/MarxistTraining/index.ts
@@ -1,5 +1,5 @@
-import KPD from '..';
-import { ZetkinJourney } from '../../../../../src/types/zetkin';
+import KPD from '../..';
+import { ZetkinJourney } from '../../../../../../src/types/zetkin';
 
 const MarxistTraining: ZetkinJourney = {
   id: 2,
